requirements.txt
numpy==1.24.2
pandas==1.5.3
requests==2.28.2
selenium==4.8.0
selenium-wire==5.1.0
Appium-Python-Client


